requirements.txt
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
backoff==2.2.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.12
exceptiongroup==1.2.2
fastapi==0.115.8
flower==2.0.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.0
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
pluggy==1.5.0
prometheus-client==0.21.1
prompt-toolkit==3.0.50
psycopg2==2.9.10
pydantic==2.10.6
pydantic-core==2.27.2
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tomli==2.2.1
tornado==6.4.2
typing-extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
uvicorn[standard]==0.34.0
